[Tech debt] Consolidate addresses (#12178)
* Consolidate addresses in platform

* Consolidate vet360

* Consolidate all-claims

* Consolidate duplicate labels file

* Consolidate Yellow Ribbon states
