+++ src/applications/disability-benefits/all-claims/config/schema.js
@@ -3,7 +3,7 @@
 // This needs to be imported from lodash/fp because that's what we use in the
 // vets-json-schema repo
 import _ from 'lodash/fp';
-import { countries } from '../../../../platform/forms/address';
+import { countries, states } from 'platform/forms/address';
 
 const documentTypes526 = [
   { value: 'L015', label: 'Buddy/Lay Statement' },
@@ -96,220 +96,8 @@ const baseAddressDef = {
   properties: {
     country: {
       type: 'string',
-      enum: [
-        'Afghanistan',
-        'Albania',
-        'Algeria',
-        'Angola',
-        'Anguilla',
-        'Antigua',
-        'Antigua and Barbuda',
-        'Argentina',
-        'Armenia',
-        'Australia',
-        'Austria',
-        'Azerbaijan',
-        'Azores',
-        'Bahamas',
-        'Bahrain',
-        'Bangladesh',
-        'Barbados',
-        'Barbuda',
-        'Belarus',
-        'Belgium',
-        'Belize',
-        'Benin',
-        'Bermuda',
-        'Bhutan',
-        'Bolivia',
-        'Bosnia-Herzegovina',
-        'Botswana',
-        'Brazil',
-        'Brunei',
-        'Bulgaria',
-        'Burkina Faso',
-        'Burma',
-        'Burundi',
-        'Cambodia',
-        'Cameroon',
-        'Canada',
-        'Cape Verde',
-        'Cayman Islands',
-        'Central African Republic',
-        'Chad',
-        'Chile',
-        'China',
-        'Colombia',
-        'Comoros',
-        'Congo, Democratic Republic of',
-        "Congo, People's Republic of",
-        'Costa Rica',
-        "Cote d'Ivoire",
-        'Croatia',
-        'Cuba',
-        'Cyprus',
-        'Czech Republic',
-        'Denmark',
-        'Djibouti',
-        'Dominica',
-        'Dominican Republic',
-        'Ecuador',
-        'Egypt',
-        'El Salvador',
-        'England',
-        'Equatorial Guinea',
-        'Eritrea',
-        'Estonia',
-        'Ethiopia',
-        'Fiji',
-        'Finland',
-        'France',
-        'French Guiana',
-        'Gabon',
-        'Gambia',
-        'Georgia',
-        'Germany',
-        'Ghana',
-        'Gibraltar',
-        'Great Britain',
-        'Great Britain and Gibraltar',
-        'Greece',
-        'Greenland',
-        'Grenada',
-        'Guadeloupe',
-        'Guatemala',
-        'Guinea',
-        'Guinea, Republic of Guinea',
-        'Guinea-Bissau',
-        'Guyana',
-        'Haiti',
-        'Honduras',
-        'Hong Kong',
-        'Hungary',
-        'Iceland',
-        'India',
-        'Indonesia',
-        'Iran',
-        'Iraq',
-        'Ireland',
-        'Israel (Jerusalem)',
-        'Israel (Tel Aviv)',
-        'Italy',
-        'Jamaica',
-        'Japan',
-        'Jordan',
-        'Kazakhstan',
-        'Kenya',
-        'Kosovo',
-        'Kuwait',
-        'Kyrgyzstan',
-        'Laos',
-        'Latvia',
-        'Lebanon',
-        'Leeward Islands',
-        'Lesotho',
-        'Liberia',
-        'Libya',
-        'Liechtenstein',
-        'Lithuania',
-        'Luxembourg',
-        'Macao',
-        'Macedonia',
-        'Madagascar',
-        'Malawi',
-        'Malaysia',
-        'Mali',
-        'Malta',
-        'Martinique',
-        'Mauritania',
-        'Mauritius',
-        'Mexico',
-        'Moldavia',
-        'Mongolia',
-        'Montenegro',
-        'Montserrat',
-        'Morocco',
-        'Mozambique',
-        'Namibia',
-        'Nepal',
-        'Netherlands',
-        'Netherlands Antilles',
-        'Nevis',
-        'New Caledonia',
-        'New Zealand',
-        'Nicaragua',
-        'Niger',
-        'Nigeria',
-        'North Korea',
-        'Northern Ireland',
-        'Norway',
-        'Oman',
-        'Pakistan',
-        'Panama',
-        'Papua New Guinea',
-        'Paraguay',
-        'Peru',
-        'Philippines',
-        'Philippines (restricted payments)',
-        'Poland',
-        'Portugal',
-        'Qatar',
-        'Republic of Yemen',
-        'Romania',
-        'Russia',
-        'Rwanda',
-        'Sao-Tome/Principe',
-        'Saudi Arabia',
-        'Scotland',
-        'Senegal',
-        'Serbia',
-        'Serbia/Montenegro',
-        'Seychelles',
-        'Sicily',
-        'Sierra Leone',
-        'Singapore',
-        'Slovakia',
-        'Slovenia',
-        'Somalia',
-        'South Africa',
-        'South Korea',
-        'Spain',
-        'Sri Lanka',
-        'St. Kitts',
-        'St. Lucia',
-        'St. Vincent',
-        'Sudan',
-        'Suriname',
-        'Swaziland',
-        'Sweden',
-        'Switzerland',
-        'Syria',
-        'Taiwan',
-        'Tajikistan',
-        'Tanzania',
-        'Thailand',
-        'Togo',
-        'Trinidad and Tobago',
-        'Tunisia',
-        'Turkey (Adana only)',
-        'Turkey (except Adana)',
-        'Turkmenistan',
-        'USA',
-        'Uganda',
-        'Ukraine',
-        'United Arab Emirates',
-        'United Kingdom',
-        'Uruguay',
-        'Uzbekistan',
-        'Vanuatu',
-        'Venezuela',
-        'Vietnam',
-        'Wales',
-        'Western Samoa',
-        'Yemen Arab Republic',
-        'Zambia',
-        'Zimbabwe',
-      ],
+      enum: countries.map(country => country.value),
+      enumNames: countries.map(country => country.label),
       default: 'USA',
     },
     addressLine1: {
@@ -334,138 +122,8 @@ const baseAddressDef = {
     },
     state: {
       type: 'string',
-      enum: [
-        'AL',
-        'UM',
-        'AS',
-        'AZ',
-        'AR',
-        'AA',
-        'AE',
-        'AP',
-        'CA',
-        'CO',
-        'CT',
-        'DE',
-        'DC',
-        'FM',
-        'FL',
-        'GA',
-        'GU',
-        'HI',
-        'ID',
-        'IL',
-        'IN',
-        'IA',
-        'KS',
-        'KY',
-        'LA',
-        'ME',
-        'MH',
-        'MD',
-        'MA',
-        'MI',
-        'MN',
-        'MS',
-        'AK',
-        'MT',
-        'NE',
-        'NV',
-        'NH',
-        'NJ',
-        'NM',
-        'NY',
-        'NC',
-        'ND',
-        'MP',
-        'OH',
-        'OK',
-        'OR',
-        'PW',
-        'PA',
-        'PR',
-        'RI',
-        'SC',
-        'SD',
-        'TN',
-        'TX',
-        'UT',
-        'VT',
-        'VI',
-        'VA',
-        'WA',
-        'WV',
-        'WI',
-        'WY',
-        'PI',
-        'MO',
-      ],
-      enumNames: [
-        'Alabama',
-        'U.S. Minor Outlying Islands',
-        'American Samoa',
-        'Arizona',
-        'Arkansas',
-        'Armed Forces Americas (AA)',
-        'Armed Forces Europe (AE)',
-        'Armed Forces Pacific (AP)',
-        'California',
-        'Colorado',
-        'Connecticut',
-        'Delaware',
-        'District Of Columbia',
-        'Federated States Of Micronesia',
-        'Florida',
-        'Georgia',
-        'Guam',
-        'Hawaii',
-        'Idaho',
-        'Illinois',
-        'Indiana',
-        'Iowa',
-        'Kansas',
-        'Kentucky',
-        'Louisiana',
-        'Maine',
-        'Marshall Islands',
-        'Maryland',
-        'Massachusetts',
-        'Michigan',
-        'Minnesota',
-        'Mississippi',
-        'Alaska',
-        'Montana',
-        'Nebraska',
-        'Nevada',
-        'New Hampshire',
-        'New Jersey',
-        'New Mexico',
-        'New York',
-        'North Carolina',
-        'North Dakota',
-        'Northern Mariana Islands',
-        'Ohio',
-        'Oklahoma',
-        'Oregon',
-        'Palau',
-        'Pennsylvania',
-        'Puerto Rico',
-        'Rhode Island',
-        'South Carolina',
-        'South Dakota',
-        'Tennessee',
-        'Texas',
-        'Utah',
-        'Vermont',
-        'Virgin Islands',
-        'Virginia',
-        'Washington',
-        'West Virginia',
-        'Wisconsin',
-        'Wyoming',
-        'Philippine Islands',
-        'Missouri',
-      ],
+      enum: states.USA.map(state => state.value),
+      enumNames: states.USA.map(state => state.label),
     },
     zipCode: {
       type: 'string',
