--- src/applications/vic-v2/labels.js
@@ -1,860 +1,0 @@
-export const branchCodeLabels = {
-  A: 'Army',
-  C: 'Coast Guard',
-  F: 'Air Force',
-  N: 'Navy',
-  M: 'Marine Corps',
-};
-
-export const countryValues = [
-  'AFG',
-  'ALA',
-  'ALB',
-  'DZA',
-  'AND',
-  'AGO',
-  'AIA',
-  'ATA',
-  'ATG',
-  'ARG',
-  'ARM',
-  'ABW',
-  'AUS',
-  'AUT',
-  'AZE',
-  'BHS',
-  'BHR',
-  'BGD',
-  'BRB',
-  'BLR',
-  'BEL',
-  'BLZ',
-  'BEN',
-  'BMU',
-  'BTN',
-  'BOL',
-  'BES',
-  'BIH',
-  'BWA',
-  'BVT',
-  'BRA',
-  'IOT',
-  'BRN',
-  'BGR',
-  'BFA',
-  'BDI',
-  'KHM',
-  'CMR',
-  'CAN',
-  'CPV',
-  'CYM',
-  'CAF',
-  'TCD',
-  'CHL',
-  'CHN',
-  'TWN',
-  'CXR',
-  'CCK',
-  'COL',
-  'COM',
-  'COG',
-  'COD',
-  'COK',
-  'CRI',
-  'CIV',
-  'HRV',
-  'CUB',
-  'CUW',
-  'CYP',
-  'CZE',
-  'DNK',
-  'DJI',
-  'DMA',
-  'DOM',
-  'ECU',
-  'EGY',
-  'SLV',
-  'GNQ',
-  'ERI',
-  'EST',
-  'ETH',
-  'FLK',
-  'FRO',
-  'FJI',
-  'FIN',
-  'FRA',
-  'GUF',
-  'PYF',
-  'ATF',
-  'GAB',
-  'GMB',
-  'GEO',
-  'DEU',
-  'GHA',
-  'GIB',
-  'GRC',
-  'GRL',
-  'GRD',
-  'GLP',
-  'GTM',
-  'GGY',
-  'GIN',
-  'GNB',
-  'GUY',
-  'HTI',
-  'HMD',
-  'VAT',
-  'HND',
-  'HUN',
-  'ISL',
-  'IND',
-  'IDN',
-  'IRN',
-  'IRQ',
-  'IRL',
-  'IMN',
-  'ISR',
-  'ITA',
-  'JAM',
-  'JPN',
-  'JEY',
-  'JOR',
-  'KAZ',
-  'KEN',
-  'KIR',
-  'PRK',
-  'KOR',
-  'KWT',
-  'KGZ',
-  'LAO',
-  'LVA',
-  'LBN',
-  'LSO',
-  'LBR',
-  'LBY',
-  'LIE',
-  'LTU',
-  'LUX',
-  'MAC',
-  'MKD',
-  'MDG',
-  'MWI',
-  'MYS',
-  'MDV',
-  'MLI',
-  'MLT',
-  'MTQ',
-  'MRT',
-  'MUS',
-  'MYT',
-  'MEX',
-  'MDA',
-  'MCO',
-  'MNG',
-  'MNE',
-  'MSR',
-  'MAR',
-  'MOZ',
-  'MMR',
-  'NAM',
-  'NRU',
-  'NPL',
-  'NLD',
-  'NCL',
-  'NZL',
-  'NIC',
-  'NER',
-  'NGA',
-  'NIU',
-  'NFK',
-  'NOR',
-  'OMN',
-  'PAK',
-  'PSE',
-  'PAN',
-  'PNG',
-  'PRY',
-  'PER',
-  'PHL',
-  'PCN',
-  'POL',
-  'PRT',
-  'QAT',
-  'REU',
-  'ROU',
-  'RUS',
-  'RWA',
-  'BLM',
-  'SHN',
-  'KNA',
-  'LCA',
-  'MAF',
-  'SPM',
-  'VCT',
-  'WSM',
-  'SMR',
-  'STP',
-  'SAU',
-  'SEN',
-  'SRB',
-  'SYC',
-  'SLE',
-  'SGP',
-  'SXM',
-  'SVK',
-  'SVN',
-  'SLB',
-  'SOM',
-  'ZAF',
-  'SGS',
-  'SSD',
-  'ESP',
-  'LKA',
-  'SDN',
-  'SUR',
-  'SJM',
-  'SWZ',
-  'SWE',
-  'CHE',
-  'SYR',
-  'TJK',
-  'TZA',
-  'THA',
-  'TLS',
-  'TGO',
-  'TKL',
-  'TON',
-  'TTO',
-  'TUN',
-  'TUR',
-  'TKM',
-  'TCA',
-  'TUV',
-  'UGA',
-  'UKR',
-  'ARE',
-  'GBR',
-  'USA',
-  'URY',
-  'UZB',
-  'VUT',
-  'VEN',
-  'VNM',
-  'VGB',
-  'WLF',
-  'ESH',
-  'YEM',
-  'ZMB',
-  'ZWE',
-];
-
-export const countryLabels = [
-  'Afghanistan',
-  'Aland Islands',
-  'Albania',
-  'Algeria',
-  'Andorra',
-  'Angola',
-  'Anguilla',
-  'Antarctica',
-  'Antigua and Barbuda',
-  'Argentina',
-  'Armenia',
-  'Aruba',
-  'Australia',
-  'Austria',
-  'Azerbaijan',
-  'Bahamas',
-  'Bahrain',
-  'Bangladesh',
-  'Barbados',
-  'Belarus',
-  'Belgium',
-  'Belize',
-  'Benin',
-  'Bermuda',
-  'Bhutan',
-  'Bolivia, Plurinational State of',
-  'Bonaire, Sint Eustatius and Saba',
-  'Bosnia and Herzegovina',
-  'Botswana',
-  'Bouvet Island',
-  'Brazil',
-  'British Indian Ocean Territory',
-  'Brunei Darussalam',
-  'Bulgaria',
-  'Burkina Faso',
-  'Burundi',
-  'Cambodia',
-  'Cameroon',
-  'Canada',
-  'Cape Verde',
-  'Cayman Islands',
-  'Central African Republic',
-  'Chad',
-  'Chile',
-  'China',
-  'Chinese Taipei',
-  'Christmas Island',
-  'Cocos (Keeling) Islands',
-  'Colombia',
-  'Comoros',
-  'Congo',
-  'Congo, the Democratic Republic of the',
-  'Cook Islands',
-  'Costa Rica',
-  "Cote d'Ivoire",
-  'Croatia',
-  'Cuba',
-  'Curaçao',
-  'Cyprus',
-  'Czech Republic',
-  'Denmark',
-  'Djibouti',
-  'Dominica',
-  'Dominican Republic',
-  'Ecuador',
-  'Egypt',
-  'El Salvador',
-  'Equatorial Guinea',
-  'Eritrea',
-  'Estonia',
-  'Ethiopia',
-  'Falkland Islands (Malvinas)',
-  'Faroe Islands',
-  'Fiji',
-  'Finland',
-  'France',
-  'French Guiana',
-  'French Polynesia',
-  'French Southern Territories',
-  'Gabon',
-  'Gambia',
-  'Georgia',
-  'Germany',
-  'Ghana',
-  'Gibraltar',
-  'Greece',
-  'Greenland',
-  'Grenada',
-  'Guadeloupe',
-  'Guatemala',
-  'Guernsey',
-  'Guinea',
-  'Guinea-Bissau',
-  'Guyana',
-  'Haiti',
-  'Heard Island and McDonald Islands',
-  'Holy See (Vatican City State)',
-  'Honduras',
-  'Hungary',
-  'Iceland',
-  'India',
-  'Indonesia',
-  'Iran, Islamic Republic of',
-  'Iraq',
-  'Ireland',
-  'Isle of Man',
-  'Israel',
-  'Italy',
-  'Jamaica',
-  'Japan',
-  'Jersey',
-  'Jordan',
-  'Kazakhstan',
-  'Kenya',
-  'Kiribati',
-  "Korea, Democratic People's Republic of",
-  'Korea, Republic of',
-  'Kuwait',
-  'Kyrgyzstan',
-  "Lao People's Democratic Republic",
-  'Latvia',
-  'Lebanon',
-  'Lesotho',
-  'Liberia',
-  'Libyan Arab Jamahiriya',
-  'Liechtenstein',
-  'Lithuania',
-  'Luxembourg',
-  'Macao',
-  'Macedonia, the former Yugoslav Republic of',
-  'Madagascar',
-  'Malawi',
-  'Malaysia',
-  'Maldives',
-  'Mali',
-  'Malta',
-  'Martinique',
-  'Mauritania',
-  'Mauritius',
-  'Mayotte',
-  'Mexico',
-  'Moldova, Republic of',
-  'Monaco',
-  'Mongolia',
-  'Montenegro',
-  'Montserrat',
-  'Morocco',
-  'Mozambique',
-  'Myanmar',
-  'Namibia',
-  'Nauru',
-  'Nepal',
-  'Netherlands',
-  'New Caledonia',
-  'New Zealand',
-  'Nicaragua',
-  'Niger',
-  'Nigeria',
-  'Niue',
-  'Norfolk Island',
-  'Norway',
-  'Oman',
-  'Pakistan',
-  'Palestinian Territory, Occupied',
-  'Panama',
-  'Papua New Guinea',
-  'Paraguay',
-  'Peru',
-  'Philippines',
-  'Pitcairn',
-  'Poland',
-  'Portugal',
-  'Qatar',
-  'Reunion',
-  'Romania',
-  'Russian Federation',
-  'Rwanda',
-  'Saint Barthélemy',
-  'Saint Helena, Ascension and Tristan da Cunha',
-  'Saint Kitts and Nevis',
-  'Saint Lucia',
-  'Saint Martin (French part)',
-  'Saint Pierre and Miquelon',
-  'Saint Vincent and the Grenadines',
-  'Samoa',
-  'San Marino',
-  'Sao Tome and Principe',
-  'Saudi Arabia',
-  'Senegal',
-  'Serbia',
-  'Seychelles',
-  'Sierra Leone',
-  'Singapore',
-  'Sint Maarten (Dutch part)',
-  'Slovakia',
-  'Slovenia',
-  'Solomon Islands',
-  'Somalia',
-  'South Africa',
-  'South Georgia and the South Sandwich Islands',
-  'South Sudan',
-  'Spain',
-  'Sri Lanka',
-  'Sudan',
-  'Suriname',
-  'Svalbard and Jan Mayen',
-  'Swaziland',
-  'Sweden',
-  'Switzerland',
-  'Syrian Arab Republic',
-  'Tajikistan',
-  'Tanzania, United Republic of',
-  'Thailand',
-  'Timor-Leste',
-  'Togo',
-  'Tokelau',
-  'Tonga',
-  'Trinidad and Tobago',
-  'Tunisia',
-  'Turkey',
-  'Turkmenistan',
-  'Turks and Caicos Islands',
-  'Tuvalu',
-  'Uganda',
-  'Ukraine',
-  'United Arab Emirates',
-  'United Kingdom',
-  'United States',
-  'Uruguay',
-  'Uzbekistan',
-  'Vanuatu',
-  'Venezuela, Bolivarian Republic of',
-  'Viet Nam',
-  'Virgin Islands, British',
-  'Wallis and Futuna',
-  'Western Sahara',
-  'Yemen',
-  'Zambia',
-  'Zimbabwe',
-];
-
-export const states = {
-  BRA: [
-    'Acre',
-    'Alagoas',
-    'Amapá',
-    'Amazonas',
-    'Bahia',
-    'Ceará',
-    'Distrito Federal',
-    'Espírito Santo',
-    'Goiás',
-    'Maranhão',
-    'Mato Grosso',
-    'Mato Grosso do Sul',
-    'Minas Gerais',
-    'Pará',
-    'Paraíba',
-    'Paraná',
-    'Pernambuco',
-    'Piauí',
-    'Rio de Janeiro',
-    'Rio Grande do Norte',
-    'Rio Grande do Sul',
-    'Rondônia',
-    'Roraima',
-    'Santa Catarina',
-    'São Paulo',
-    'Sergipe',
-    'Tocantins',
-  ],
-  ITA: [
-    'Agrigento',
-    'Alessandria',
-    'Ancona',
-    'Aosta',
-    'Arezzo',
-    'Ascoli Piceno',
-    'Asti',
-    'Avellino',
-    'Bari',
-    'Barletta-Andria-Trani',
-    'Belluno',
-    'Benevento',
-    'Bergamo',
-    'Biella',
-    'Bologna',
-    'Bolzano',
-    'Brescia',
-    'Brindisi',
-    'Cagliari',
-    'Caltanissetta',
-    'Campobasso',
-    'Carbonia-Iglesias',
-    'Caserta',
-    'Catania',
-    'Catanzaro',
-    'Chieti',
-    'Como',
-    'Cosenza',
-    'Cremona',
-    'Crotone',
-    'Cuneo',
-    'Enna',
-    'Fermo',
-    'Ferrara',
-    'Florence',
-    'Foggia',
-    'Forlì-Cesena',
-    'Frosinone',
-    'Genoa',
-    'Gorizia',
-    'Grosseto',
-    'Imperia',
-    'Isernia',
-    "L'Aquila",
-    'La Spezia',
-    'Latina',
-    'Lecce',
-    'Lecco',
-    'Livorno',
-    'Lodi',
-    'Lucca',
-    'Macerata',
-    'Mantua',
-    'Massa and Carrara',
-    'Matera',
-    'Medio Campidano',
-    'Messina',
-    'Milan',
-    'Modena',
-    'Monza and Brianza',
-    'Naples',
-    'Novara',
-    'Nuoro',
-    'Ogliastra',
-    'Olbia-Tempio',
-    'Oristano',
-    'Padua',
-    'Palermo',
-    'Parma',
-    'Pavia',
-    'Perugia',
-    'Pesaro and Urbino',
-    'Pescara',
-    'Piacenza',
-    'Pisa',
-    'Pistoia',
-    'Pordenone',
-    'Potenza',
-    'Prato',
-    'Ragusa',
-    'Ravenna',
-    'Reggio Calabria',
-    'Reggio Emilia',
-    'Rieti',
-    'Rimini',
-    'Rome',
-    'Rovigo',
-    'Salerno',
-    'Sassari',
-    'Savona',
-    'Siena',
-    'Sondrio',
-    'Syracuse',
-    'Taranto',
-    'Teramo',
-    'Terni',
-    'Trapani',
-    'Trento',
-    'Treviso',
-    'Trieste',
-    'Turin',
-    'Udine',
-    'Varese',
-    'Venice',
-    'Verbano-Cusio-Ossola',
-    'Vercelli',
-    'Verona',
-    'Vibo Valentia',
-    'Vicenza',
-    'Viterbo',
-  ],
-  MEX: [
-    'Aguascalientes',
-    'Baja California',
-    'Baja California Sur',
-    'Campeche',
-    'Chiapas',
-    'Chihuahua',
-    'Coahuila',
-    'Colima',
-    'Durango',
-    'Federal District',
-    'Guanajuato',
-    'Guerrero',
-    'Hidalgo',
-    'Jalisco',
-    'Mexico State',
-    'Michoacán',
-    'Morelos',
-    'Nayarit',
-    'Nuevo León',
-    'Oaxaca',
-    'Puebla',
-    'Querétaro',
-    'Quintana Roo',
-    'San Luis Potosí',
-    'Sinaloa',
-    'Sonora',
-    'Tabasco',
-    'Tamaulipas',
-    'Tlaxcala',
-    'Veracruz',
-    'Yucatán',
-    'Zacatecas',
-  ],
-  USA: [
-    'Armed Forces Americas',
-    'Armed Forces Europe',
-    'Alaska',
-    'Alabama',
-    'Armed Forces Pacific',
-    'Arkansas',
-    'American Samoa',
-    'Arizona',
-    'California',
-    'Colorado',
-    'Connecticut',
-    'District of Columbia',
-    'Delaware',
-    'Florida',
-    'Federated Micronesia',
-    'Georgia',
-    'Guam',
-    'Hawaii',
-    'Iowa',
-    'Idaho',
-    'Illinois',
-    'Indiana',
-    'Kansas',
-    'Kentucky',
-    'Louisiana',
-    'Massachusetts',
-    'Maryland',
-    'Maine',
-    'Marshall Islands',
-    'Michigan',
-    'Minnesota',
-    'Missouri',
-    'Northern Mariana Islands',
-    'Mississippi',
-    'Montana',
-    'North Carolina',
-    'North Dakota',
-    'Nebraska',
-    'New Hampshire',
-    'New Jersey',
-    'New Mexico',
-    'Nevada',
-    'New York',
-    'Ohio',
-    'Oklahoma',
-    'Oregon',
-    'Pennsylvania',
-    'Puerto Rico',
-    'Palau',
-    'Rhode Island',
-    'South Carolina',
-    'South Dakota',
-    'Tennessee',
-    'Texas',
-    'United States Minor Outlying Islands',
-    'Utah',
-    'Virginia',
-    'US Virgin Islands',
-    'Vermont',
-    'Washington',
-    'Wisconsin',
-    'West Virginia',
-    'Wyoming',
-  ],
-  CAN: [
-    'Alberta',
-    'British Columbia',
-    'Manitoba',
-    'New Brunswick',
-    'Newfoundland and Labrador',
-    'Northwest Territories',
-    'Nova Scotia',
-    'Nunavut',
-    'Ontario',
-    'Prince Edward Island',
-    'Quebec',
-    'Saskatchewan',
-    'Yukon Territories',
-  ],
-  IND: [
-    'Andaman and Nicobar Islands',
-    'Andhra Pradesh',
-    'Arunachal Pradesh',
-    'Assam',
-    'Bihar',
-    'Chandigarh',
-    'Chhattisgarh',
-    'Dadra and Nagar Haveli',
-    'Daman and Diu',
-    'Delhi',
-    'Goa',
-    'Gujarat',
-    'Haryana',
-    'Himachal Pradesh',
-    'Jammu and Kashmir',
-    'Jharkhand',
-    'Karnataka',
-    'Kerala',
-    'Lakshadweep',
-    'Madhya Pradesh',
-    'Maharashtra',
-    'Manipur',
-    'Meghalaya',
-    'Mizoram',
-    'Nagaland',
-    'Odisha',
-    'Puducherry',
-    'Punjab',
-    'Rajasthan',
-    'Sikkim',
-    'Tamil Nadu',
-    'Tripura',
-    'Uttarakhand',
-    'Uttar Pradesh',
-    'West Bengal',
-  ],
-  CHN: [
-    'Anhui',
-    'Beijing',
-    'Chinese Taipei',
-    'Chongqing',
-    'Fujian',
-    'Gansu',
-    'Guangdong',
-    'Guangxi',
-    'Guizhou',
-    'Hainan',
-    'Hebei',
-    'Heilongjiang',
-    'Henan',
-    'Hong Kong',
-    'Hubei',
-    'Hunan',
-    'Jiangsu',
-    'Jiangxi',
-    'Jilin',
-    'Liaoning',
-    'Macao',
-    'Nei Mongol',
-    'Ningxia',
-    'Qinghai',
-    'Shaanxi',
-    'Shandong',
-    'Shanghai',
-    'Shanxi',
-    'Sichuan',
-    'Tianjin',
-    'Xinjiang',
-    'Xizang',
-    'Yunnan',
-    'Zhejiang',
-  ],
-  AUS: [
-    'Australian Capital Territory',
-    'New South Wales',
-    'Northern Territory',
-    'Queensland',
-    'South Australia',
-    'Tasmania',
-    'Victoria',
-    'Western Australia',
-  ],
-  IRL: [
-    'Carlow',
-    'Cavan',
-    'Clare',
-    'Cork',
-    'Donegal',
-    'Dublin',
-    'Galway',
-    'Kerry',
-    'Kildare',
-    'Kilkenny',
-    'Laois',
-    'Leitrim',
-    'Limerick',
-    'Longford',
-    'Louth',
-    'Mayo',
-    'Meath',
-    'Monaghan',
-    'Offaly',
-    'Roscommon',
-    'Sligo',
-    'Tipperary',
-    'Waterford',
-    'Westmeath',
-    'Wexford',
-    'Wicklow',
-  ],
-};
