--- src/platform/forms/address/data/military-cities.json
@@ -1,5 +1,0 @@
-[
-  "APO",
-  "FPO",
-  "DPO"
-]
