--- src/platform/forms/address/data/military-states.json
@@ -1,5 +1,0 @@
-[
-  "AA",
-  "AE",
-  "AP"
-]
